Flood fill now supports a selection region. Added shortcuts for tools. Also added a toggle layer visibility command (default shortcut is the number 0 key)

+++ PicNetStudio.Avalonia/PicNet/Layers/MatrixUtils.cs
@@ -22,17 +22,17 @@ using SkiaSharp;
 namespace PicNetStudio.Avalonia.PicNet.Layers;
 
 public static class MatrixUtils {
-    public static SKMatrix CreateTransformationMatrix(SKPoint pos, SKPoint scale, double rotation, SKPoint scaleOrigin, SKPoint rotationOrigin) {
+    public static SKMatrix CreateTransformationMatrix(SKPoint pos, SKPoint scale, float rotation, SKPoint scaleOrigin, SKPoint rotationOrigin) {
         SKMatrix matrix = SKMatrix.Identity;
         matrix = matrix.PreConcat(SKMatrix.CreateTranslation(pos.X, pos.Y));
         matrix = matrix.PreConcat(SKMatrix.CreateScale(scale.X, scale.Y, scaleOrigin.X, scaleOrigin.Y));
-        matrix = matrix.PreConcat(SKMatrix.CreateRotationDegrees((float) rotation, rotationOrigin.X, rotationOrigin.Y));
+        matrix = matrix.PreConcat(SKMatrix.CreateRotationDegrees(rotation, rotationOrigin.X, rotationOrigin.Y));
         return matrix;
     }
 
-    public static SKMatrix CreateInverseTransformationMatrix(SKPoint pos, SKPoint scale, double rotation, SKPoint scaleOrigin, SKPoint rotationOrigin) {
+    public static SKMatrix CreateInverseTransformationMatrix(SKPoint pos, SKPoint scale, float rotation, SKPoint scaleOrigin, SKPoint rotationOrigin) {
         SKMatrix matrix = SKMatrix.Identity;
-        matrix = matrix.PreConcat(SKMatrix.CreateRotationDegrees((float) -rotation, rotationOrigin.X, rotationOrigin.Y));
+        matrix = matrix.PreConcat(SKMatrix.CreateRotationDegrees(-rotation, rotationOrigin.X, rotationOrigin.Y));
         matrix = matrix.PreConcat(SKMatrix.CreateScale(scale.X == 0.0F ? float.MaxValue : (1 / scale.X), scale.Y == 0.0F ? float.MaxValue : (1 / scale.Y), scaleOrigin.X, scaleOrigin.Y));
         matrix = matrix.PreConcat(SKMatrix.CreateTranslation(-pos.X, -pos.Y));
         return matrix;
